feat/created search by token feature
- added a page to acces a form to search an exam by the token
- moved the nav bar links into list items
- added token params in tests backend endpoint
- made tests fot this endpont with params contexts
- created an API endpoint in the frontend service. This endpoint is responsible
  for making a request to the tests backendo endpoint to get the data;
- moved the table HTML node creation to AbstractView.js class.
  SearchView and ExamsViews class inherits from it.
- created requests tests for the frontend data endpoint;
