+++ app/frontend/src/views/index.erb
@@ -3,13 +3,16 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Página Inicial - Exames Médicos</title>
+  <title>Rebase Labs 2024 - Turma 11</title>
 </head>
 <body>
   <header>
     <nav class="nav">
-      <a href="/" class="nav__link" data-link>Início</a>
-      <a href="/exams" class="nav__link" data-link>Exames</a>
+      <ul>
+        <li><a href="/" class="nav__link" data-link>Início</a></li>
+        <li><a href="/exams" class="nav__link" data-link>Exames</a></li>
+        <li><a href="/search" class="nav__link" data-link>Busca por Token</a></li>
+      </ul>
     </nav>
   </header>
   <main id="app"></main>
